pyproject.toml
[tool.poetry]
name = "boa-vvm"
version = "0.1.0"
description = ""
authors = ["Alberto <albicento.ac@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
titanoboa = "0.2.0"
vvm = "^0.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
